requirements.txt
streamlit>=1.27.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.24.1
google-generativeai>=0.3.0